@@ -77,78 +77,69 @@
         "arrowParens": "avoid"
     },
     "dependencies": {
-        "@angular/animations": "^7.2.9",
-        "@angular/common": "^7.2.9",
-        "@angular/compiler": "^7.2.9",
-        "@angular/core": "^7.2.9",
-        "@angular/forms": "^7.2.9",
-        "@angular/http": "^7.2.9",
-        "@angular/platform-browser": "^7.2.9",
-        "@angular/platform-browser-dynamic": "^7.2.9",
-        "@angular/platform-server": "^7.2.9",
-        "@angular/router": "^7.2.9",
-        "@fortawesome/angular-fontawesome": "^0.3.0",
-        "@fortawesome/fontawesome-svg-core": "^1.2.15",
-        "@fortawesome/free-solid-svg-icons": "^5.7.2",
-        "@ng-bootstrap/ng-bootstrap": "^4.1.0",
+        "@angular/animations": "^8.0.2",
+        "@angular/common": "^8.0.2",
+        "@angular/compiler": "^8.0.2",
+        "@angular/core": "^8.0.2",
+        "@angular/forms": "^8.0.2",
+        "@angular/platform-browser": "^8.0.2",
+        "@angular/platform-browser-dynamic": "^8.0.2",
+        "@angular/platform-server": "^8.0.2",
+        "@angular/router": "^8.0.2",
+        "@fortawesome/angular-fontawesome": "^0.4.0",
+        "@fortawesome/fontawesome-svg-core": "^1.2.19",
+        "@fortawesome/free-solid-svg-icons": "^5.9.0",
+        "@ng-bootstrap/ng-bootstrap": "^4.2.1",
         "bootstrap": "^4.3.1",
         "core-js": "^2.5.4",
-        "font-awesome": "^4.7.0",
-        "json2typescript": "^1.2.0",
-        "ngx-gallery": "^5.9.1",
+        "json2typescript": "^1.2.3",
+        "ngx-gallery": "^5.10.0",
         "ngx-json-viewer": "^2.4.0",
-        "rxjs": "~6.3.3",
+        "rxjs": "~6.5.2",
         "snapsvg": "^0.5.1",
         "tslib": "^1.9.0",
-        "zone.js": "~0.8.26"
+        "zone.js": "~0.9.1"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "~0.13.0",
-        "@angular/cli": "~7.3.6",
-        "@angular/compiler-cli": "^7.2.9",
-        "@angular/language-service": "^7.2.9",
-        "@commitlint/cli": "^7.5.2",
-        "@commitlint/config-angular": "^7.5.0",
-        "@types/jasmine": "~2.8.8",
+        "@angular-devkit/build-angular": "~0.800.3",
+        "@angular/cli": "~8.0.3",
+        "@angular/compiler-cli": "^8.0.2",
+        "@angular/language-service": "^8.0.2",
+        "@commitlint/cli": "^8.0.0",
+        "@commitlint/config-angular": "^8.0.0",
+        "@types/jasmine": "~3.3.8",
         "@types/jasminewd2": "~2.0.3",
         "@types/node": "~8.9.4",
         "angular-cli-ghpages": "^0.5.3",
-        "codelyzer": "~4.5.0",
-        "conventional-recommended-bump": "^4.0.4",
+        "codelyzer": "^5.0.0",
+        "conventional-recommended-bump": "^5.0.0",
         "cross-var": "^1.1.0",
-        "gzipper": "^2.2.0",
-        "husky": "^1.3.1",
-        "jasmine-core": "~2.99.1",
+        "gzipper": "^2.8.0",
+        "husky": "^2.4.1",
+        "jasmine-core": "~3.4.0",
         "jasmine-spec-reporter": "~4.2.1",
-        "karma": "~4.0.0",
+        "karma": "~4.1.0",
         "karma-chrome-launcher": "~2.2.0",
         "karma-coverage-istanbul-reporter": "~2.0.1",
-        "karma-jasmine": "~1.1.2",
-        "karma-jasmine-html-reporter": "^0.2.2",
-        "node-sass": "^4.11.0",
-        "nodemon": "^1.18.10",
+        "karma-jasmine": "~2.0.1",
+        "karma-jasmine-html-reporter": "^1.4.0",
+        "node-sass": "^4.12.0",
+        "nodemon": "^1.19.1",
         "npm-run-all": "^4.1.5",
-        "prettier": "1.16.4",
-        "pretty-quick": "^1.10.0",
+        "prettier": "1.18.2",
+        "pretty-quick": "^1.11.1",
         "protractor": "~5.4.0",
-        "source-map-explorer": "^1.8.0",
-        "standard-version": "^5.0.2",
+        "source-map-explorer": "^2.0.1",
+        "standard-version": "^6.0.1",
         "ts-node": "~7.0.0",
-        "tslint": "~5.11.0",
+        "tslint": "~5.15.0",
         "tslint-config-prettier": "^1.18.0",
-        "typescript": "~3.2.4",
-        "webpack-bundle-analyzer": "^3.1.0"
+        "typescript": "~3.4.5",
+        "webpack-bundle-analyzer": "^3.3.2"
     },
     "engines": {
         "node": ">= 8.9.0",
         "npm": ">= 5.5.1",
         "yarn": "^1.10.0"
-    },
-    "browserslist": [
-        "> 0.5%",
-        "last 2 versions",
-        "Firefox ESR",
-        "not dead",
-        "IE 9-11"
-    ]
+    }
 }
