@@ -47,7 +47,7 @@ export class ApiServiceResult {
      * @throws
      */
 
-    getBody(classObject?: { new (): any }): any {
+    getBody(classObject?: new () => any): any {
         // console.log(this.body);
         return ApiServiceResult.jsonConvert.deserialize(this.body, classObject);
     }
