@@ -20,7 +20,7 @@ const dataRoutes: Routes = [
                     { path: 'timeline', component: TimelineComponent },
                     { path: 'detail/:id', redirectTo: 'resource/:id' }, // absolute redirect (replacement of route) to resource/:id,
                     // TODO: lazy loaded bibliography path muted for now
-                    // { path: 'bibliography', loadChildren: './data-outlets/bibliography/bibliography.module#BibliographyModule'}
+                    // { path: 'bibliography', loadChildren: () => import('./data-outlets/bibliography/bibliography.module').then(m => m.BibliographyModule)}
                     { path: '', pathMatch: 'full', redirectTo: 'fulltext' }
                 ]
             }
